feat: refactor FastAPI integration and update endpoint paths for consistency
