+++ fun_report1/src/__init__.py
@@ -1,0 +1,1 @@
+"""Source package for fun_report1 Azure Function app."""
